🔀 :: 남은 날짜 계산 안되는 이슈 해결

+++ crowdfunding/src/main/kotlin/com/project/indistraw/crowdfunding/adapter/input/data/response/CrowdfundingDetailResponse.kt
@@ -8,7 +8,7 @@ data class CrowdfundingDetailResponse(
     val description: String,
     val writer: Writer,
     val amount: Amount,
-    val remainingDay: Int,
+    val remainingDay: Long,
     val fundingCount: Long,
     val reward: List<RewardResponse>,
     val status: StatusType,
